feat: move stuff to cogs
